+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpHandlerDiagnosticListener.cs
@@ -153,7 +153,7 @@ internal sealed class HttpHandlerDiagnosticListener : ListenerHandler
                 return;
             }
 
-            activity.DisplayName = HttpTagHelper.GetOperationNameForHttpMethod(request.Method);
+            RequestMethodHelper.SetHttpClientActivityDisplayName(activity, request.Method.Method);
 
             if (!IsNet7OrGreater)
             {
@@ -162,17 +162,7 @@ internal sealed class HttpHandlerDiagnosticListener : ListenerHandler
             }
 
             // see the spec https://github.com/open-telemetry/semantic-conventions/blob/v1.23.0/docs/http/http-spans.md
-            if (RequestMethodHelper.KnownMethods.TryGetValue(request.Method.Method, out var httpMethod))
-            {
-                activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, httpMethod);
-            }
-            else
-            {
-                // Set to default "_OTHER" as per spec.
-                // https://github.com/open-telemetry/semantic-conventions/blob/v1.22.0/docs/http/http-spans.md#common-attributes
-                activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, "_OTHER");
-                activity.SetTag(SemanticConventions.AttributeHttpRequestMethodOriginal, request.Method.Method);
-            }
+            RequestMethodHelper.SetHttpMethodTag(activity, request.Method.Method);
 
             activity.SetTag(SemanticConventions.AttributeServerAddress, request.RequestUri.Host);
             if (!request.RequestUri.IsDefaultPort)